Add new parameters to the Sensor struct.
A new parameter to read and keep the header of a given file. The
constructor was changed to accept a bool indicating whether a file has
a header or not. The parameter is an array of strings (char **) and a
size parameter to save its size.

Add the necessary implementation, change slightly the implementation of
sensor_get_next_scene, add appropriate test cases and update previous
ones. Passes all unit and memory leak checks.