Add nerd_utils.c, and update all building scripts.
Change int_vector.c to nerd_utils, and externalise the _trim function
from literal.c in the utils to become trim. Update all the dependencies
which includes all the headers used in the code, and the file included
in the building scripts. Update building scripts to make a memory leak
checks before exiting, and change the name of the binary that each
creates to be a variable to make it safer while re-using it in the same
script.