Update the main file.
Add two other parameters where the former requires a constraints file
(can be empty) and the latter which an optional number to be used in
the random seed. The file will create a new directory to save the
implementations parameters and the KnowledgeBase at each epoch, and it
also sets the global parameters. It passes the memory leak checks.