Add nerd_helper.c file with helpers for calling Prudens-JS.
Move prudensjs_inference function away from nerd.c to nerd_helper.c,
which aims to globalise the usage of the function. The PrudensSettings
struct (which is required) will be used for every inference call, which
will provide the necessary information for every inference call to be
successful. Update all the files and functions using this function.