Add the increasing_demotion option to rule_hypergraph_update_rules.
The new option allows to increase the demotion rate or not of the
backward chaining demotion (so far it only decreased). Add a new
parameter in the Nerd structure to keep track of it, update the
constructors and the to_file functions to accommodate the changes, and
adjust test suites. Passes all tests and memory leak checks.